pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xv6-filesystem-reader"
version = "0.1.0"
description = "Read and write xv6 file system disk images; mount them with FUSE on Linux."
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
authors = [
    { name = "xv6filesystemreader" }
]
keywords = ["xv6", "filesystem", "fuse", "disk-image"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Education",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "fusepy>=3.0",
]

[project.optional-dependencies]
dev = [
    "build>=0.10",
    "twine>=4.0",
]

[project.scripts]
xv6fs = "fusereader:main"
xv6fs-extract = "reader:main"

[tool.setuptools]
py-modules = ["reader", "fusereader"]
